refactor: move EdgeTrigger from package root into outputs/
EdgeTrigger gates a side effect to fire only when a prediction changes. Per a
Codex review: it is not an ml primitive (it knows nothing about models, features,
or pipeline state) — it shapes a *downstream effect*, so it belongs with the
send-side. outputs/ already owns "things leaving the app": outputs.filters
handles continuous vector outputs, EdgeTrigger handles discrete event outputs.

- git mv edge_trigger.py -> outputs/edge_trigger.py
- re-exported from myogestic.outputs and kept in the top-level facade, so both
  `from myogestic import EdgeTrigger` and `from myogestic.outputs import
  EdgeTrigger` work
- updated tests + docs to the facade import path

Package root now holds only the core spine (core, stream, contracts, grid) plus
private runtime support (_browser/_platform/_theme). Verified: ruff + ty +
pyright clean, 176 tests, docs 0 warnings.
